+++ core/sr-api-macros/tests/ui/changed_at_unknown_version.stderr
@@ -1,0 +1,5 @@
+error: `changed_in` version can not be greater than the `api_version`
+  --> $DIR/changed_at_unknown_version.rs:15:3
+   |
+15 |         fn test(data: u64);
+   |         ^^
